Initial attempt at porting LibJitsi to use the FIPS implementation of BouncyCastle.
Support for SKEIN, which was an optional SRTP block cipher was removed,
since it appears to be only relevant to the Jitsi SIP client's usage
of the Jitsi VideoBridge, and doesn't appear to be used in any
implementations of WebRTC.

For instance, HMAC_SHA1 is used exclusively in the WebRTC reference
implementation:
https://chromium.googlesource.com/external/webrtc/+/075af927309cea5420d06798c389c27cc6b34e61/webrtc/pc/mediasession.cc#219

The BouncyCastle Crypto libraries apparently have been removed in the
FIPS libraries, and the BC jars only include the Java Cryptography
Architecture (JCA) implementation, which wraps around BouncyCastle and
makes use of the FIPS bouncycastle implementation underneath.

It remains to be seen if this is valid, and if not we will have to
investigate if there is an implementation of the crypto bc package
somewhere.

Several types that were made package private have been exported by
making wrapper classes that are in the org.bouncycastle.fips package,
extending the package private classes, and then making them public.

This shouldn't violate the requires for FIPS validation of the module
because we do not modify the code:
https://downloads.bouncycastle.org/fips-java/BC-FJA-SecurityPolicy-1.0.0.pdf