Remove package local wrappers, and instead expose the private getProvider method in FipsRegister.
Apparently when a jar is signed, you have to sign using the same
signature to add a class to a package encompassed by that jar.

As a result it seems like using reflection to expose a single method is
a better alternative.

All Fips BlockCipherProviders, and DigestProviderss, are registered to a
single register where a singleton instance of that Cipher or Digest can
be looked up.