Connection test now succeeds with FIPS.
However, it appears SCTP (The protocol the datachannel uses) is broken
somehow. I need to investigate this more closely to determine exactly
how/why it is broken, since most of the code that was ported to support
the FIPS implementation of BouncyCastle were isolated to the DTLS/Tls
Client Server modules.

The main changes here are directly passing in an instance of a
BouncyCastleFipsProvider directly to the JcaTlsCryptoProvider, rather
than expecting it to load the provider by string, because apparently to
work this way it has to be registered inside the java.security file. As
well as modifying the default fingerprint algorithm to be sha-256, and
adjusting the names returned by DefaultAlgorithmNameFinder to include a
dash, since this gets passed directly through to the generated SDP,
which expects those algorithm names to have a dash.

I.E: sha-256 as opposed to sha256