BouncyCastle's FIPS replacement implemented a DTLSClientProtocol, which is used by LibJitsi's DtlsPacketTransformer class to mutate and modify the DTLS stream, that destroys the established masterSecret from a passed in SecurityContext object.
LibJitsi was grabbing this secret, and using it for initializing other
parts of the DtlsPacketTransformer, and when this secret was
unexpectedly null, it will fail in all sorts of other unexpected ways.

Unfortunately the behavior to clear this was nested deep inside of
DTLSClientProtocol, but the class itself was thankfully public!

So, we created a wrapper class, that replaces the connect method with a
method that does NOT destroy the master secret.

Another option might be to perhaps override the clear method on the
SecurityContext itself, or some other means, but this seemed like the
most direct method.

Unfortunately implementing it was a super pain, because a large number
of the functions and classes being used internal to connect are all
private, so we have to use reflection very heavily to accomplish our
desired outcome.

In the end the transformer works, and I was able to establish a working
media session with logging coming from the transformer regarding packets
passing through it!